server/requirements.txt
grpcio==1.59.3
grpcio-tools==1.59.3
names==0.3.0
protobuf==4.25.1
